requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
openai==1.88.0
python-dotenv==1.1.0